fixed pull request feedback; renamed to have a consintent naming
