+++ waveshare2in13v2/doc.go
@@ -2,9 +2,7 @@
 // Use of this source code is governed under the Apache License, Version 2.0
 // that can be found in the LICENSE file.
 
-// Package epd controls Waveshare e-paper series displays.
-//
-// More details
+// Package waveshare2in13v2 controls Waveshare 2in13v2 e-paper display.
 //
 // Datasheets
 //
@@ -13,5 +11,5 @@
 // Product page:
 //
 // 2.13 Inch version 2: https://www.waveshare.com/wiki/2.13inch_e-Paper_HAT
-
-package waveshare213v2
+//
+package waveshare2in13v2
